17 commit

+++ src/main/java/com/example/sviter/repository/MessageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 
-public interface MessageRepository extends CrudRepository<MyMassage, Integer> {
+public interface MessageRepository extends CrudRepository<MyMassage, Long> {
     List<MyMassage> findByTag(String tag);
 
 }
